# Conflicts:
#	prosolo-main/src/main/webapp/manage/js/quadrantchart.js

+++ prosolo-analytics/src/main/scala/org/prosolo/bigdata/scala/clustering/sna/UserNode.scala
@@ -6,7 +6,7 @@ package org.prosolo.bigdata.scala.clustering.sna
 /**
   * zoran 20/12/15
   */
-class UserNode(val id:Int) {
+class UserNode(val id:Long) {
 
   var cluster = 0
